refactor: add button folder in shared component
